cmdexec: add lvmCommand() and LVMWrapper()
With the new `HEKETI_LVM_WRAPPER` environment variable or the
`lvm_wrapper` configuration option in `heketi.json`, it is now possible
to run LVM and related commands though a wrapper. This is useful for
environments where Gluster is containerized and `nsenter` or similar is
required to execute LVM commands in a stable fashion.

Signed-off-by: Niels de Vos <ndevos@redhat.com>